requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.0
